Исправление ошибок с 7 урока(2)
