+++ app/src/main/java/space/kuz/notesapp/ui/fragment/SettingFragment.java
@@ -1,9 +1,7 @@
-package space.kuz.notesapp.fragment;
+package space.kuz.notesapp.ui.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
