+++ app/src/main/res/layout/fragment_basket.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFEB3B"
-    tools:context=".fragment.BasketFragment">
+    tools:context=".ui.fragment.BasketFragment">
 
 
     <EditText
