+++ app/src/main/res/layout/fragment_person.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#0040FF"
-    tools:context=".fragment.PersonFragment">
+    tools:context=".ui.fragment.PersonFragment">
 
     <EditText
         android:id="@+id/fragment_person_edit_text"
